Modified database metadata to store archetype_id to help with auth
